+++ CHANGES.md
@@ -5,6 +5,8 @@ Unreleased
 -   Drop support for Python 3.9.
 -   Add `get_or_abort` and `one_or_abort` methods, which get a single row or
     otherwise tell Flask to abort with a 404 error.
+-   Add `test_isolation` context manager, which isolates changes to the database
+    so that tests don't affect each other.
 
 ## Version 0.1.0